requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
pyee==9.0.4
python-dotenv==0.21.0
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
yarl==1.8.2
